refactor axios endpoint
