+++ camunda-formio-tasklist-vue/src/services/axios.ts
@@ -11,3 +11,30 @@ export function bpmAxios (bearerToken: string, CamundaUrl: any) {
     },
   });
 }
+
+export const httpPOSTRequest = (
+  url: string,
+  data: any,
+  bearerToken: string,
+  isBearer = true
+) => {
+  return axios.post(url, data, {
+    headers: {
+      Authorization: isBearer ? `Bearer ${bearerToken}` : bearerToken,
+    },
+  });
+};
+
+export const httpGETRequest = (
+  url: string,
+  data: any,
+  bearerToken: string,
+  isBearer = true
+) => {
+  return axios.get(url, {
+    params: data,
+    headers: {
+      Authorization: isBearer ? `Bearer ${bearerToken}` : bearerToken,
+    },
+  });
+};
