+++ camunda-formio-tasklist-vue/src/services/formsflowai-api.ts
@@ -1,31 +1,4 @@
-import axios from "axios";
-
-export const httpPOSTRequest = (
-  url: string,
-  data: any,
-  token: string,
-  isBearer = true
-) => {
-  return axios.post(url, data, {
-    headers: {
-      Authorization: isBearer ? `Bearer ${token}` : token,
-    },
-  });
-};
-
-export const httpGETRequest = (
-  url: string,
-  data: any,
-  token: string,
-  isBearer = true
-) => {
-  return axios.get(url, {
-    params: data,
-    headers: {
-      Authorization: isBearer ? `Bearer ${token}` : token,
-    },
-  });
-};
+import { httpGETRequest, httpPOSTRequest } from "./axios";
 
 export const formApplicationSubmit = (
   Apiurl: string,
